refactor(oauth): Update README and tools for unified Google services authentication
This commit enhances the documentation and refactors the OAuth tools for Google services, consolidating the authentication process for Gmail, Google Docs, and Google Drive. Key changes include:

- Updated README to reflect the unified OAuth token usage across Google services.
- Removed redundant OAuth tools specific to Gmail and Docs, replacing them with a single `google_save_auth_code` tool for authentication.
- Added new tools for Google OAuth flow, streamlining the user experience.
- Improved error handling and documentation for the authentication process.

These changes simplify the integration of Google services and provide a consistent user experience across the application.
