+++ internal/tools/docs_tools/tools.go
@@ -14,28 +14,6 @@ import (
 
 // RegisterDocsTools registers all Google Docs-related tools with the MCP server
 func RegisterDocsTools(s *mcpserver.MCPServer, sc *server.ServerContext) error {
-	// Get OAuth URL tool
-	getAuthURLTool := mcp.NewTool("docs_get_auth_url",
-		mcp.WithDescription("Get the OAuth URL to authorize Google services access (Gmail, Docs, Drive)"),
-	)
-
-	s.AddTool(getAuthURLTool, func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		return handleGetAuthURL(ctx, request, sc)
-	})
-
-	// Save authorization code tool
-	saveAuthCodeTool := mcp.NewTool("docs_save_auth_code",
-		mcp.WithDescription("Save the OAuth authorization code to complete Google services authentication (Gmail, Docs, Drive)"),
-		mcp.WithString("authCode",
-			mcp.Required(),
-			mcp.Description("The authorization code from Google OAuth"),
-		),
-	)
-
-	s.AddTool(saveAuthCodeTool, func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		return handleSaveAuthCode(ctx, request, sc)
-	})
-
 	// Get document tool
 	getDocumentTool := mcp.NewTool("docs_get_document",
 		mcp.WithDescription("Get Google Docs content by document ID"),
@@ -97,7 +75,7 @@ func handleGetDocument(ctx context.Context, request mcp.CallToolRequest, sc *ser
 4. Copy the authorization code
 
 5. Provide the authorization code to your AI agent
-   The agent will use the docs_save_auth_code or gmail_save_auth_code tool to complete authentication.
+   The agent will use the google_save_auth_code tool to complete authentication.
 
 Note: You only need to authorize once. The tokens will be automatically refreshed.`, authURL)
 			return mcp.NewToolResultError(errorMsg), nil
@@ -169,7 +147,7 @@ func handleGetMetadata(ctx context.Context, request mcp.CallToolRequest, sc *ser
 4. Copy the authorization code
 
 5. Provide the authorization code to your AI agent
-   The agent will use the docs_save_auth_code or gmail_save_auth_code tool to complete authentication.
+   The agent will use the google_save_auth_code tool to complete authentication.
 
 Note: You only need to authorize once. The tokens will be automatically refreshed.`, authURL)
 			return mcp.NewToolResultError(errorMsg), nil
@@ -195,37 +173,4 @@ Note: You only need to authorize once. The tokens will be automatically refreshe
 
 	result := fmt.Sprintf("Document metadata:\n%s", string(jsonBytes))
 	return mcp.NewToolResultText(result), nil
-}
-
-func handleGetAuthURL(ctx context.Context, request mcp.CallToolRequest, sc *server.ServerContext) (*mcp.CallToolResult, error) {
-	authURL := docs.GetAuthURL()
-
-	result := fmt.Sprintf(`To authorize Google services access (Gmail, Docs, Drive):
-
-1. Visit this URL in your browser:
-   %s
-
-2. Sign in with your Google account
-3. Grant access to Google services
-4. Copy the authorization code
-
-5. Call the docs_save_auth_code or gmail_save_auth_code tool with the code to complete authentication`, authURL)
-
-	return mcp.NewToolResultText(result), nil
-}
-
-func handleSaveAuthCode(ctx context.Context, request mcp.CallToolRequest, sc *server.ServerContext) (*mcp.CallToolResult, error) {
-	args := request.GetArguments()
-
-	authCode, ok := args["authCode"].(string)
-	if !ok || authCode == "" {
-		return mcp.NewToolResultError("authCode is required"), nil
-	}
-
-	err := docs.SaveToken(ctx, authCode)
-	if err != nil {
-		return mcp.NewToolResultError(fmt.Sprintf("Failed to save authorization code: %v", err)), nil
-	}
-
-	return mcp.NewToolResultText("✅ Authorization successful! Google services token saved. You can now use all Gmail and Google Docs tools."), nil
 }
